update of the 3d cta simulator
